Recepies

+++ Homie.API/DTOs/Ingredient/IngredientResponseBody.cs
@@ -5,15 +5,15 @@
         public Guid Id { get; set; }
         public string IngredientName { get; set; }
         public string? Description { get; set; }
-        public float? Calories { get; set; }
-        public float? Carbs { get; set; }
-        public float? Fat { get; set; }
-        public float? Protein { get; set; }
-        public float? ServingAmount { get; set; }
+        public double? Calories { get; set; }
+        public double? Carbs { get; set; }
+        public double? Fat { get; set; }
+        public double? Protein { get; set; }
+        public double? ServingAmount { get; set; }
         public string? ServingDescription { get; set; }
-        public float? GramsPerServing { get; set; }
+        public double? GramsPerServing { get; set; }
 
-        public IngredientResponseBody(Guid id, string ingredientName, string? description, float? calories, float? carbs, float? fat, float? protein, float? servingAmount, string? servingDescription, float? gramsPerServing)
+        public IngredientResponseBody(Guid id, string ingredientName, string? description, double? calories, double? carbs, double? fat, double? protein, double? servingAmount, string? servingDescription, double? gramsPerServing)
         {
             Id = id;
             IngredientName = ingredientName;
