+++ Homie.API/DTOs/Ingredient/UpdateIngredientRequestBody.cs
@@ -4,15 +4,15 @@
     {
         public string FoodName { get; set; }
         public string? Description { get; set; }
-        public float? Calories { get; set; }
-        public float? Carbs { get; set; }
-        public float? Fat { get; set; }
-        public float? Protein { get; set; }
-        public float? ServingAmount { get; set; }
+        public double? Calories { get; set; }
+        public double? Carbs { get; set; }
+        public double? Fat { get; set; }
+        public double? Protein { get; set; }
+        public double? ServingAmount { get; set; }
         public string? ServingDescription { get; set; }
-        public float? GramsPerServing { get; set; }
+        public double? GramsPerServing { get; set; }
 
-        public UpdateIngredientRequestBody(string foodName, string? description, float? calories, float? carbs, float? fat, float? protein, float? servingAmount, string? servingDescription, float? gramsPerServing)
+        public UpdateIngredientRequestBody(string foodName, string? description, double? calories, double? carbs, double? fat, double? protein, double? servingAmount, string? servingDescription, double? gramsPerServing)
         {
             FoodName = foodName;
             Description = description;
