+++ Homie.API/Models/Food.cs
@@ -7,7 +7,7 @@
         public string? Description { get; set; }
         public Guid? PlaceId { get; set; }
         public Guid? IngredientId { get; set; }
-        public float? IngredientAmount { get; set; }
+        public double? IngredientAmount { get; set; }
         public DateTime? ExpirationDate { get; set; }
         public DateTime? LastUpdated { get; set; }
     }
